разбил MainWindow по отдельным partial классам
