+++ DrawMoar/DrawMoar.csproj
@@ -105,6 +105,11 @@
     <Compile Include="ILayer.cs" />
     <Compile Include="Instrument.cs" />
     <Compile Include="Drawing\ffmpeg\ImportImage.cs" />
+    <Compile Include="MainWin.Buttons.cs" />
+    <Compile Include="MainWin.CanvasEvents.cs" />
+    <Compile Include="MainWin.Frames.cs" />
+    <Compile Include="MainWin.Layers.cs" />
+    <Compile Include="MainWin.Scenes.cs" />
     <Compile Include="MainWindow.xaml.cs">
       <DependentUpon>MainWindow.xaml</DependentUpon>
     </Compile>
