+++ DrawMoar/MainWindow.xaml
@@ -49,8 +49,8 @@
         <WrapPanel Name="figurePanel" Grid.Column="0" Grid.Row="1" MaxWidth="130" Grid.RowSpan="2">
             <WrapPanel Name="wp1" Width="130" Height="180">
                 <Button x:Name="Lines" Height ="25" Width="40" Margin="5,5" Content="Lines" Click="Lines_Click"/>
-                <Button x:Name="testButton" Height ="25" Width="40" Margin="5,5" Content="Brush" Click="testButton_Click"/>
-                <Button x:Name="testButton2" Height ="25" Width="40" Margin="5,5" Content="Arrow" Click="testButton2_Click"/>
+                <Button x:Name="Brush" Height ="25" Width="40" Margin="5,5" Content="Brush" Click="Brush_Click"/>
+                <Button x:Name="Arrow" Height ="25" Width="40" Margin="5,5" Content="Arrow" Click="Arrow_Click"/>
                 <Button x:Name="AddEllipse" Height ="25" Width="40" Margin="5,5" Content="Ellipse" Click="AddEllipse_Click"/>
                 <Button x:Name="AddRectangle" Height ="25" Width="60" Margin="5,5" Content="Rectangle" Click="AddRectangle_Click"/>
                 <Button x:Name="AddAudioButton" Height ="25" Width="60" Margin="5,5" Content="Audio" Click="AddAudio"/>
@@ -59,9 +59,9 @@
             </WrapPanel>
             <WrapPanel Name="wp2" Width="130" MinHeight="90" Height="135">
                 <Button x:Name="CycleFrame" Content="Cycle" Width="40"  Height="25" Margin="5,2" Click="CycleFrame_Click"/>
-                <Button x:Name="GenerateFrame" Content="Generate" Width="25"  Height="23" Margin="5,2" Click="GenerateFrame_Click"/>
-                <Button x:Name="AddTransformation" Content="Add Transform" Width="25"  Height="24" Margin="5,2" Click="AT_Click"/>
-                <Button x:Name="AFrameTrans" Content="Add Frame Transform" Width="34"  Height="22" Margin="5,2" Click="AFT_Click"/>
+                <Button x:Name="GenerateFrame" Content="Generate" Width="auto"  Height="23" Margin="5,2" Click="GenerateFrame_Click"/>
+                <Button x:Name="AddTransformation" Content="Transform" Width="auto"  Height="24" Margin="5,2" Click="AT_Click"/>
+                <Button x:Name="AFrameTrans" Content="Frame Transform" Width="auto"  Height="22" Margin="5,2" Click="AFT_Click"/>
                 <xctk:ColorPicker x:Name="ClrPcker_Background"  Margin="70,3,0,0" Height="25" Width="45" SelectedColor="Black" SelectedColorChanged="ClrPcker_Background_SelectedColorChanged"/>
                 <Slider x:Name="slider" Width="130" Minimum="3" Maximum="50"/>
             </WrapPanel>
