+++ DrawMoar/MainWindow.xaml.cs
@@ -76,37 +76,7 @@ namespace DrawMoar
         }
 
 
-        private void ExportToMP4(object sender, RoutedEventArgs e) {        
-        }
-
-
-        private void AddRasterLayer_Click(object sender, RoutedEventArgs e) {
-            if (cartoon == null) {
-                return;
-            }
-            //проверка на то, выделен ли какой-либо кадр(когда реализуем удаление)
-            if (sender != null) {
-                GlobalState.CurrentFrame.layers.Add(new Tuple<ILayer, List<Transformation>, int>(new RasterLayer(), new List<Transformation>(), 0));
-            }
-            GlobalState.CurrentLayer = GlobalState.CurrentFrame.layers.Last();
-            var layers = GlobalState.CurrentFrame.layers;
-            AddListBoxElement(layersList, $"RasterLayer_{layers.Count - 1}");
-        }
-
-
-        private void AddVectorLayer_Click(object sender, RoutedEventArgs e) {
-            if (cartoon == null) {
-                return;
-            }
-            if (sender == null) {
-                layersList.Items.Clear();
-            }
-            else {
-                GlobalState.CurrentFrame.layers.Add(new Tuple<ILayer, List<Transformation>, int>(new VectorLayer(), new List<Transformation>(), 0));
-            }
-            GlobalState.CurrentLayer = GlobalState.CurrentFrame.layers.Last();
-            var layers = GlobalState.CurrentFrame.layers;
-            AddListBoxElement(layersList, $"VectorLayer_{layers.Count - 1}");
+        private void ExportToMP4(object sender, RoutedEventArgs e) {
         }
 
 
@@ -127,216 +97,11 @@ namespace DrawMoar
         }
 
 
-        private void framesList_SelectionChanged(object sender, SelectionChangedEventArgs e) {
-            if (framesList.SelectedIndex != -1) {
-                if (GlobalState.CurrentFrame.layers.Count > 0)
-                GlobalState.CurrentFrame = GlobalState.CurrentScene.frames[framesList.SelectedIndex];
-                GlobalState.CurrentLayer = GlobalState.CurrentFrame.layers.Last();
-            }
-            layersList.Items.Clear();
-            canvas.Children.Clear();
-            var lays = GlobalState.CurrentFrame.layers;
-            foreach (var item in lays) {
-                AddListBoxElement(layersList, item.Item1.Name);
-                item.Item1.Draw(canvasDrawer);
-                layersList.SelectedIndex = 0;
-            }
-        }
-
-
-        private void scenesList_SelectionChanged(object sender, SelectionChangedEventArgs e) {
-            framesList.Items.Clear();
-            if (scenesList.SelectedIndex != -1)
-                GlobalState.CurrentScene = cartoon.scenes[scenesList.SelectedIndex];
-            var frames = GlobalState.CurrentScene.frames;
-            foreach (var item in frames) {
-                AddListBoxElement(framesList, item.Name);
-            }
-            framesList.SelectedIndex = 0;
-        }
-
-
-        private void layersList_SelectionChanged(object sender, SelectionChangedEventArgs e) {
-            if (layersList.SelectedIndex != -1) {
-                GlobalState.CurrentLayer = GlobalState.CurrentFrame.layers[layersList.SelectedIndex];
-            }
-
-        }
-
-
-        private void testButton_Click(object sender, RoutedEventArgs e) {
-            GlobalState.CurrentTool = Instrument.Brush;
-        }
-
-
-        private void testButton2_Click(object sender, RoutedEventArgs e) {
-            GlobalState.CurrentTool = Instrument.Arrow;
-        }
-
-
-        private void AddFrame_Click(object sender, RoutedEventArgs e) {
-            if (cartoon == null) {
-                return;
-            }
-            if (sender != null) {
-                GlobalState.CurrentScene.frames.Add(new BaseElements.Frame());
-                GlobalState.CurrentFrame = GlobalState.CurrentScene.frames.Last();
-                GlobalState.CurrentLayer = GlobalState.CurrentFrame.layers.Last();
-                var frames = GlobalState.CurrentScene.frames;
-                AddListBoxElement(framesList, GlobalState.CurrentFrame.Name);
-            }
-        }
-
-
-        private void AddScene_Click(object sender, RoutedEventArgs e) {
-            if (cartoon == null) {
-                return;
-            }
-            if (sender != null) {
-                cartoon.scenes.Add(new Scene());
-            }
-            GlobalState.CurrentScene = cartoon.scenes.Last();
-            GlobalState.CurrentFrame = GlobalState.CurrentScene.frames.Last();
-            GlobalState.CurrentLayer = GlobalState.CurrentFrame.layers.Last();
-            AddListBoxElement(scenesList, GlobalState.CurrentScene.Name);
-        }
-
-
         private void AddListBoxElement(ListBox lBox, string content) {
             var lbl = new Label();          
             lbl.Content = content;
             lBox.Items.Add(lbl);
             lBox.SelectedIndex = lBox.Items.Count - 1;
-        }
-
-
-        void canvas_MouseLeftButtonDown(object sender, MouseButtonEventArgs e) {
-            GlobalState.PressLeftButton = true;
-            GlobalState.BrushSize = new Size(slider.Value, slider.Value);
-            var currentLayer = GlobalState.CurrentLayer.Item1;
-            prevPoint = Mouse.GetPosition(canvas);
-            switch (GlobalState.CurrentTool) {
-                case Instrument.Arrow:
-                    break;
-                case Instrument.Brush:
-                    break;
-                case Instrument.Rectangle:
-                    newRect = new Rectangle(prevPoint, new Size(15, 10));
-                    newRect.Draw(canvasDrawer);
-                    SaveIntoLayer(currentLayer, newRect);
-                    break;
-                case Instrument.Ellipse:
-                    newEllipse = new Ellipse(prevPoint, new Size(15, 10));
-                    newEllipse.Draw(canvasDrawer);
-                    SaveIntoLayer(currentLayer, newEllipse);
-                    break;
-                case Instrument.Line:
-                    newLine = new Line(prevPoint, prevPoint);
-                    newLine.Draw(canvasDrawer);
-                    SaveIntoLayer(currentLayer, newLine);
-                    break;
-                    case Instrument.Eraser:
-                    break;
-            }
-            canvas_MouseMove(sender, e);
-        }
-
-
-        void canvas_MouseMove(object sender, MouseEventArgs e) {
-            point = (Point)e.GetPosition(canvas);
-            var currentLayer = GlobalState.CurrentLayer.Item1;
-            if (!GlobalState.PressLeftButton) return;
-            switch (GlobalState.CurrentTool) {
-                case Instrument.Arrow:
-                    TranslatingRedrawing(e);
-                    prevPoint = point;
-                    break;
-                case Instrument.Brush:
-                    newLine = new Line(prevPoint, point);
-                    newLine.Draw(canvasDrawer);
-                    SaveIntoLayer(currentLayer, newLine);                   
-                    prevPoint = point;
-                    break;
-                case Instrument.Rectangle:
-                    ScaleRedrawing(newRect, e);
-                    break;
-                case Instrument.Ellipse:
-                    ScaleRedrawing(newEllipse, e);
-                    break;
-                case Instrument.Line:
-                    ScaleRedrawing(newLine, e);
-                    break;
-                case Instrument.Eraser:
-                    var bufColor = ClrPcker_Background.SelectedColor.Value;
-                    ClrPcker_Background.SelectedColor = Colors.Transparent;
-                    newLine = new Line(prevPoint, point);
-                    newLine.Draw(canvasDrawer);
-                    SaveIntoLayer(currentLayer, newLine);
-                    prevPoint = point;
-                    ClrPcker_Background.SelectedColor = bufColor;
-                    break;
-            }
-        }
-
-
-        void ScaleRedrawing(IShape shape, MouseEventArgs e) {
-            if (shape != null & e.LeftButton == MouseButtonState.Pressed) {
-                var layer = GlobalState.CurrentLayer;
-                var shiftX = point.X - prevPoint.X;
-                var shiftY = point.Y - prevPoint.Y;
-                if (((shiftX <= 0) || (shiftY <= 0)) && (!(shape is Line))) return;   //пока из-за "плохих" шифтов так; уберу, когда сделаю зеркалирование
-                canvas.Children.RemoveAt(canvas.Children.Count - 1);
-
-                if (shape is Line) shape = new Line(prevPoint, point);
-                else if (shape is Ellipse) shape = new Ellipse(prevPoint, new Size(15 + shiftX, 10 + shiftY));
-                else if (shape is Rectangle) shape = new Rectangle(prevPoint, new Size(15 + shiftX, 10 + shiftY));
-                shape.Draw(canvasDrawer);
-                if (layer.Item1 is VectorLayer) {
-                    var shapes = ((VectorLayer)layer.Item1).Picture.shapes;
-                    shapes.RemoveAt(shapes.Count - 1);
-                    SaveIntoLayer(layer.Item1, shape);
-                }
-                //else ((RasterLayer)layer.Item1).Save(canvas);
-            }
-        }
-
-
-        void TranslatingRedrawing(MouseEventArgs e) {
-            point = e.GetPosition(canvas);
-            GlobalState.CurrentLayer.Item1.Transform(new TranslateTransformation(new Point(point.X - prevPoint.X, point.Y - prevPoint.Y)));
-            Refresh();
-        }
-
-
-        void SaveIntoLayer(ILayer layer, IShape shape) {
-            if (layer is VectorLayer) {
-                ((VectorLayer)layer).Picture.shapes.Add(shape);
-            }
-        }
-
-
-        void canvas_MouseLeftButtonUp(object sender, MouseButtonEventArgs e) {
-            GlobalState.PressLeftButton = false;
-        }
-
-
-        void canvas_MouseLeave(object sender, MouseEventArgs e) {
-            GlobalState.PressLeftButton = false;
-        }
-
-
-        private void Lines_Click(object sender, RoutedEventArgs e) {
-            GlobalState.CurrentTool = Instrument.Line;
-        }
-
-
-        private void AddEllipse_Click(object sender, RoutedEventArgs e) {
-            GlobalState.CurrentTool = Instrument.Ellipse;
-        }
-
-
-        private void AddRectangle_Click(object sender, RoutedEventArgs e) {
-            GlobalState.CurrentTool = Instrument.Rectangle;
         }
 
 
@@ -358,15 +123,6 @@ namespace DrawMoar
 
         private void Window_SizeChanged(object sender, SizeChangedEventArgs e) {
             menu.Width = Width;
-        }
-
-
-        private void GenerateFrame_Click(object sender, RoutedEventArgs e) {
-            if (cartoon == null) return;
-            if (GlobalState.TotalTime == 0) return;
-            GlobalState.CurrentScene.Generate(GlobalState.CurrentFrame, GlobalState.TotalTime);
-            scenesList_SelectionChanged(null, null);
-            Refresh();
         }
 
 
@@ -396,58 +152,6 @@ namespace DrawMoar
         }
 
 
-        private void DeleteFrame_Click(object sender, RoutedEventArgs e) {
-            if (cartoon == null) return;
-            int index = framesList.SelectedIndex;
-            framesList.Items.RemoveAt(index);
-            var frames = GlobalState.CurrentScene.frames;
-            frames.RemoveAt(index);
-            if (frames.Count == 0) {
-                frames.Add(new BaseElements.Frame());
-                AddListBoxElement(framesList, GlobalState.CurrentFrame.Name);
-            }
-            framesList.SelectedIndex = index > 0 ? index - 1 : 0;
-            GlobalState.CurrentFrame = index > 0 ? frames[index - 1] : frames[0];
-            GlobalState.CurrentLayer = GlobalState.CurrentFrame.layers[0];
-            Refresh();
-        }
-
-
-        private void DeleteLayer_Click(object sender, RoutedEventArgs e) {
-            if (cartoon == null) return;
-            int index = layersList.SelectedIndex;
-            var layerToDelete = GlobalState.CurrentFrame.layers[index].Item1;
-            layersList.Items.RemoveAt(index);
-            var layers = GlobalState.CurrentFrame.layers;
-            layers.RemoveAt(index);
-            if (layers.Count == 0) {
-                layers.Add(new Tuple<ILayer, List<Transformation>, int>(new VectorLayer(), new List<Transformation>(), 0));
-                AddListBoxElement(layersList, GlobalState.CurrentLayer.Item1.Name);
-            }
-            layersList.SelectedIndex = layersList.Items.Count > 1 ? index - 1 : 0;
-            GlobalState.CurrentLayer = index > 0 ? layers[index - 1] : layers[0];
-            Refresh();
-        }
-        
-
-        //кнопку запилю с ПК, когда пофикшу штуку с гитом, что скидывал в конфу
-        private void DeleteScene_Click(object sender, RoutedEventArgs e) {  //вероятно, возможно вынести по классам и часть вызывать оттуда
-            if (cartoon == null) return;
-            int index = scenesList.SelectedIndex;
-            scenesList.Items.RemoveAt(index);
-            cartoon.scenes.RemoveAt(index);
-            if (cartoon.scenes.Count == 0) {
-                cartoon.scenes.Add(new Scene());
-                AddListBoxElement(scenesList, GlobalState.CurrentScene.Name);
-            }
-            scenesList.SelectedIndex = index > 0 ? index - 1 : 0;
-            GlobalState.CurrentScene = index > 0 ? cartoon.scenes[index - 1] : cartoon.scenes[0];
-            GlobalState.CurrentFrame = GlobalState.CurrentScene.frames[0];
-            GlobalState.CurrentLayer = GlobalState.CurrentFrame.layers[0];
-            Refresh();
-        }
-
-
         private void Refresh() {
             canvas.Children.Clear();
             var layers = GlobalState.CurrentFrame.layers;
@@ -457,24 +161,9 @@ namespace DrawMoar
         }
 
 
-        private void AT_Click(object sender, RoutedEventArgs e) {
-            ILayer cloneOfCurrent = (ILayer)GlobalState.CurrentLayer.Item1.Clone();
-            generationWin = new GenerationDialog(cloneOfCurrent);
-            generationWin.Show();
-        }
-
-
         private void Window_Closed(object sender, EventArgs e) {
             if (generationWin != null) generationWin.Close();
         }
-
-
-        private void CycleFrame_Click(object sender, RoutedEventArgs e) {
-            GlobalState.CurrentScene.Cycle(25);
-            scenesList_SelectionChanged(null, null);
-            Refresh();
-        }
-
 
         private void AFT_Click(object sender, RoutedEventArgs e) {
             
@@ -494,11 +183,6 @@ namespace DrawMoar
             var pathToAudio = d.FileName;
             audio.Text = pathToAudio;
             cartoon.pathToAudio = pathToAudio;
-        }
-
-
-        private void Eraser_Click(object sender, RoutedEventArgs e) {
-            GlobalState.CurrentTool = Instrument.Eraser;
         }
     }
 }
